fix: put theme and season into data store
we need to update those depending on the month of the question
in the other store it owuld not update
