+++ src/Answers.svelte
@@ -3,9 +3,9 @@
   import { chunkArray } from "./utils";
   import Table from "./TableV2.svelte";
   import ThermometerIcon from "./icons/Thermometer.icon.svelte";
-  import { answer, result, currentTheme } from "./stores/game.store";
+  import { answer, result } from "./stores/game.store";
 
-  import { question, data, otherAnswers } from "./stores/data.store";
+  import { question, data, otherAnswers, currentTheme } from "./stores/data.store";
   export let columns = [
     { key: "island", name: "Island" },
     { key: "month", name: "Month" },
